Improved structure by moving/renaming classes and removing unused/deprecated methods
