+++ src/Enums/DatabaseTransactionType.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Weap\Junction\Enums;
+
+enum DatabaseTransactionType
+{
+    case Store;
+    case Update;
+    case Destroy;
+    case Action;
+}
