--- src/Http/Controllers/Enums/DatabaseTransactionTypeEnum.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Weap\Junction\Http\Controllers\Enums;
-
-enum DatabaseTransactionTypeEnum
-{
-    case Store;
-    case Update;
-    case Destroy;
-    case Action;
-}
