+++ src/Http/Requests/DefaultFormRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Weap\Junction\Http\Controllers\Requests;
+namespace Weap\Junction\Http\Requests;
 
 use Illuminate\Contracts\Validation\Validator;
 use Illuminate\Foundation\Http\FormRequest;
