+++ src/Http/Resources/BaseResource.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Weap\Junction\Http\Controllers\Resources;
+namespace Weap\Junction\Http\Resources;
 
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
